version 3 for official server
version 2 for official server

improved ships needed calculation, +1 turn for neutral expansion

ships_available_safe() added - considers possible attack of enemy upon leaving

choose safe ships selection only when we're winning

bugfix: 757fc16 follow-up

bugfix: 5f670d6 follow-up, ally/enemy simmetry

config: fallback weren't printed in verbose mode

bugfix 5f670d6 follow-up, use predicted score at previous turn

target shuffle: added random permutations of enemy's targets

safe ships selection is now flag-based

dirty hack to simulate enemy move with both safe and unsafe ships selection

bugfix: fd59a39 follow-up, really use the flag now

configs updated

krysa-8 sumbission, unifying configs

expansion targets cutting off

safe ships departure refactored

simple_closeness reversed, for cutoff() usage

cutoff is done by simple_closeness

evacuation initial implementation

safe ships departure refactored

exact_attack(): safe ships enabled too

sneak defence is always unsafe

configs updated

krysa-9 submission: unifying configs

enabling windows build

a couple of type conversion warnings fixed

Conflicts:

	attacks.cpp

fastexpand: initialize initial vector properly

windows warning fix

attacks.cpp assertion fix

windows-specific timer fix

improved timeouts handling

safe/unsafe enemy move selection refactored

+++ move.h
@@ -30,7 +30,7 @@ public:
     void add(const Move *move);
     const vector<FutureOrder> & orders() const;
 
-    uint16_t size() const {return _orders.size();}
+    uint16_t size() const {return (uint16_t)_orders.size();}
     bool empty() const {return _orders.empty();}
 
     vector<FutureOrder> & mod_orders() {return _orders;}
